teams-busylight/requirements.txt
busylight-for-humans==0.33.3
sdnotify==0.3.2
